backend/requirements.txt
Flask==2.3.2
flask-pymongo==2.3.0
pymongo==4.3.3
python-dotenv==1.1.1
flask-cors==4.0.1
gunicorn==22.0.0
dnspython==2.6.1
certifi==2024.7.4
bcrypt==4.1.3
cffi>=1.15